# lib/api-spec/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Api",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/claims/start": {
      "post": {
        "operationId": "startClaim",
        "tags": [
          "ecomstack"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/onboarding/complete": {
      "post": {
        "operationId": "completeOnboarding",
        "tags": [
          "ecomstack"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "401": {
            "description": "Sign in required"
          }
        }
      }
    },
    "/claims/check": {
      "post": {
        "operationId": "checkClaim",
        "tags": [
          "ecomstack"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
...
    "/admin/impact/sync": {
      "post": {
        "operationId": "runImpactSync",
        "tags": [
          "ecomstack"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStatus"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
...
    "/admin/assets/upload": {
      "post": {
        "operationId": "requestAssetUpload",
        "tags": [
          "ecomstack"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/admin/assets/confirm": {
      "post": {
        "operationId": "confirmAsset",
        "tags": [
          "ecomstack"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetConfirm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/admin/assets/{id}": {
      "delete": {
        "operationId": "deleteAsset",
        "tags": [
          "ecomstack"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/assets/{id}/download": {
      "get": {
        "operationId": "downloadAsset",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized file stream",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Message": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "Asset": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "resourceId": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "contentType",
          "size",
          "resourceId",
          "kind"
        ]
      },
...
      "ClaimInput": {
        "type": "object",
        "properties": {
          "resourceSlug": {
            "type": "string"
          }
        },
        "required": []
      },
      "ClaimResult": {
        "type": "object",
        "properties": {
          "redirectUrl": {
            "type": "string"
          },
          "claimStatus": {
            "type": "string"
          }
        },
        "required": [
          "redirectUrl",
          "claimStatus"
        ]
      },
...
      "DiagnosticRecord": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string"
          },
          "campaignName": {
            "type": "string"
          },
          "eventTrackerId": {
            "type": "string"
          },
          "eventTrackerName": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "hasSubId1": {
            "type": "boolean"
          },
          "hasSubId2": {
            "type": "boolean"
          },
          "hasSubId3": {
            "type": "boolean"
          },
          "hasSharedId": {
            "type": "boolean"
          }
        },
        "required": [
          "campaignId",
          "campaignName",
          "eventTrackerId",
          "eventTrackerName",
          "state",
          "hasSubId1",
          "hasSubId2",
          "hasSubId3",
          "hasSharedId"
        ]
      },
      "SyncStatus": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "credentialsPresent": {
            "type": "boolean"
          },
          "lastCheckedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastSuccessAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "nextCheckAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticRecord"
            }
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "enabled",
          "credentialsPresent",
          "lastCheckedAt",
          "lastSuccessAt",
          "nextCheckAt",
          "error",
          "records",
          "message"
        ]
      },
      "UploadInput": {
        "type": "object",
        "properties": {
          "resourceId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "contentType": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "file",
              "cover"
            ]
          }
        },
        "required": [
          "resourceId",
          "name",
          "size",
          "contentType",
          "kind"
        ]
      },
      "UploadResult": {
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string"
          },
          "uploadURL": {
            "type": "string"
          }
        },
        "required": [
          "assetId",
          "uploadURL"
        ]
      },
      "AssetConfirm": {
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string"
          }
        },
        "required": [
          "assetId"
        ]
      },
...
    }
  }
}